Move cacert.h out of deps/, since it does not belong there.
